Try to refactor codes not to return final output String, but to return just required data
Try to add struct that stores result data.
Try to call the struct initializer in perform- functions of AdminMode and UserMode class to return only data required to print.
But, 'Command failed due to signal: Illegal instruction: 4' error occured.
It's because of the new made struct (Result struct).
Have to remove the new made Result struct.